# docs/openapi.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "Mem0 API Docs",
		"description": "mem0.ai API Docs",
		"contact": {
			"email": "deshraj@mem0.ai"
		},
		"license": {
			"name": "Apache 2.0"
		},
		"version": "v1"
	},
	"servers": [
		{
			"url": "https://api.mem0.ai/"
		}
	],
	"security": [
		{
			"ApiKeyAuth": []
		}
	],
	"paths": {
...
		"/v2/memories/": {
			"post": {
				"tags": [
					"memories"
				],
				"description": "Get all memories.",
				"operationId": "memories_list_v2",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MemoryGetInputV2"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successfully retrieved memories.",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"type": "object",
										"properties": {
											"id": {
												"type": "string"
											},
											"memory": {
												"type": "string"
											},
											"created_at": {
												"type": "string",
												"format": "date-time"
											},
											"updated_at": {
												"type": "string",
												"format": "date-time"
											},
											"owner": {
												"type": "string"
											},
											"immutable": {
												"type": "boolean",
												"description": "Whether the memory is immutable.",
												"title": "Immutable",
												"default": false
											},
											"expiration_date": {
												"type": "string",
												"format": "date-time",
												"description": "The date and time when the memory will expire. Format: YYYY-MM-DD.",
												"title": "Expiration date",
												"nullable": true,
												"default": null
											},
											"organization": {
												"type": "string"
											},
											"metadata": {
												"type": "object"
											}
										},
										"required": [
											"id",
											"memory",
											"created_at",
											"updated_at",
											"total_memories",
											"owner",
											"organization",
											"type"
										]
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad Request.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"example": "One of the filters: app_id, user_id, agent_id, run_id is required!"
										}
									}
								}
							}
						}
					}
				},
				"x-code-samples": [
					{
						"lang": "Python",
						"source": "# To use the Python SDK, install the package:\n# pip install mem0ai\n\nfrom mem0 import MemoryClient\nclient = MemoryClient(api_key=\"your_api_key\", org_id=\"your_org_id\", project_id=\"your_project_id\")\n\n# Retrieve memories with filters\nmemories = client.get_all(\n    filters={\n        \"AND\": [\n            {\n                \"user_id\": \"alex\"\n            },\n            {\n                \"created_at\": {\n                    \"gte\": \"2024-07-01\",\n                    \"lte\": \"2024-07-31\"\n                }\n            }\n        ]\n    },\n    version=\"v2\"\n)\n\nprint(memories)"
					},
					{
						"lang": "JavaScript",
						"source": "// To use the JavaScript SDK, install the package:\n// npm i mem0ai\n\nimport MemoryClient from 'mem0ai';\nconst client = new MemoryClient({ apiKey: \"your-api-key\" });\n\nconst filters = {\n  AND: [\n    { user_id: 'alex' },\n    { created_at: { gte: '2024-07-01', lte: '2024-07-31' } }\n  ]\n};\n\nclient.getAll({ filters, api_version: 'v2' })\n  .then(result => console.log(result))\n  .catch(error => console.error(error));"
					},
					{
						"lang": "cURL",
						"source": "curl -X POST 'https://api.mem0.ai/v2/memories/' \\\n-H 'Authorization: Token your-api-key' \\\n-H 'Content-Type: application/json' \\\n-d '{\n  \"filters\": {\n    \"AND\": [\n      { \"user_id\": \"alex\" },\n      { \"created_at\": { \"gte\": \"2024-07-01\", \"lte\": \"2024-07-31\" } }\n    ]\n  },\n  \"org_id\": \"your-org-id\",\n  \"project_id\": \"your-project-id\"\n}'"
					},
					{
						"lang": "Go",
						"source": "package main\n\nimport (\n\t\"bytes\"\n\t\"encoding/json\"\n\t\"fmt\"\n\t\"io/ioutil\"\n\t\"net/http\"\n)\n\nfunc main() {\n\turl := \"https://api.mem0.ai/v2/memories/\"\n\tfilters := map[string]interface{}{\n\t\t\"AND\": []map[string]interface{}{\n\t\t\t{\"user_id\": \"alex\"},\n\t\t\t{\"created_at\": map[string]string{\n\t\t\t\t\"gte\": \"2024-07-01\",\n\t\t\t\t\"lte\": \"2024-07-31\",\n\t\t\t}},\n\t\t},\n\t}\n\tpayload, _ := json.Marshal(map[string]interface{}{\"filters\": filters})\n\treq, _ := http.NewRequest(\"POST\", url, bytes.NewBuffer(payload))\n\treq.Header.Add(\"Authorization\", \"Token your-api-key\")\n\treq.Header.Add(\"Content-Type\", \"application/json\")\n\n\tres, _ := http.DefaultClient.Do(req)\n\tdefer res.Body.Close()\n\tbody, _ := ioutil.ReadAll(res.Body)\n\n\tfmt.Println(string(body))\n}"
					},
					{
						"lang": "PHP",
						"source": "<?php\n\n$curl = curl_init();\n\n$filters = [\n  'AND' => [\n    ['user_id' => 'alex'],\n    ['created_at' => ['gte' => '2024-07-01', 'lte' => '2024-07-31']]\n  ]\n];\n\ncurl_setopt_array($curl, [\n  CURLOPT_URL => \"https://api.mem0.ai/v2/memories/\",\n  CURLOPT_RETURNTRANSFER => true,\n  CURLOPT_ENCODING => \"\",\n  CURLOPT_MAXREDIRS => 10,\n  CURLOPT_TIMEOUT => 30,\n  CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,\n  CURLOPT_CUSTOMREQUEST => \"POST\",\n  CURLOPT_POSTFIELDS => json_encode(['filters' => $filters]),\n  CURLOPT_HTTPHEADER => [\n    \"Authorization: Token your-api-key\",\n    \"Content-Type: application/json\"\n  ],\n]);\n\n$response = curl_exec($curl);\n$err = curl_error($curl);\n\ncurl_close($curl);\n\nif ($err) {\n  echo \"cURL Error #:\" . $err;\n} else {\n  echo $response;\n}"
					},
					{
						"lang": "Java",
						"source": "import com.konghq.unirest.http.HttpResponse;\nimport com.konghq.unirest.http.Unirest;\nimport org.json.JSONObject;\n\nJSONObject filters = new JSONObject()\n    .put(\"AND\", new JSONArray()\n        .put(new JSONObject().put(\"user_id\", \"alex\"))\n        .put(new JSONObject().put(\"created_at\", new JSONObject()\n            .put(\"gte\", \"2024-07-01\")\n            .put(\"lte\", \"2024-07-31\")\n        ))\n    );\n\nHttpResponse<String> response = Unirest.post(\"https://api.mem0.ai/v2/memories/\")\n  .header(\"Authorization\", \"Token your-api-key\")\n  .header(\"Content-Type\", \"application/json\")\n  .body(new JSONObject().put(\"filters\", filters).toString())\n  .asString();\n\nSystem.out.println(response.getBody());"
					}
				]
			}
		},
...
		"/v1/memories/{entity_type}/{entity_id}/": {
			"get": {
				"tags": [
					"memories"
				],
				"operationId": "memories_read",
				"responses": {
					"200": {
						"description": "Successfully retrieved memories.",
						"content": {}
					}
				}
			},
			"parameters": [
				{
					"name": "entity_type",
					"in": "path",
					"required": true,
					"schema": {
						"type": "string"
					}
				},
				{
					"name": "entity_id",
					"in": "path",
					"required": true,
					"schema": {
						"type": "string"
					}
				}
			]
		},
...
	},
	"components": {
		"schemas": {
...
			"MemoryGetInputV2": {
				"type": "object",
				"required": [
					"filters"
				],
				"properties": {
					"filters": {
						"title": "Filters",
						"type": "object",
						"description": "A dictionary of filters to apply to retrieve memories. Available fields are: user_id, agent_id, app_id, run_id, created_at, updated_at, categories, keywords. Supports logical operators (AND, OR) and comparison operators (in, gte, lte, gt, lt, ne, contains, icontains, *). For categories field, use 'contains' for partial matching (e.g., {\"categories\": {\"contains\": \"finance\"}}) or 'in' for exact matching (e.g., {\"categories\": {\"in\": [\"personal_information\"]}}).",
						"properties": {
							"user_id": {
								"type": "string"
							},
							"agent_id": {
								"type": "string"
							},
							"app_id": {
								"type": "string"
							},
							"run_id": {
								"type": "string"
							},
							"created_at": {
								"type": "string",
								"format": "date-time"
							},
							"updated_at": {
								"type": "string",
								"format": "date-time"
							},
							"keywords": {
								"type": "object",
								"properties": {
									"contains": {
										"type": "string"
									},
									"icontains": {
										"type": "string"
									}
								}
							},
							"categories": {
								"type": "object",
								"properties": {
									"in": {
										"type": "array",
										"items": {
											"type": "string"
										}
									}
								}
							},
							"metadata": {
								"type": "object"
							}
						},
						"additionalProperties": {
							"type": "object",
							"properties": {
								"in": {
									"type": "array"
								},
								"gte": {
									"type": "string"
								},
								"lte": {
									"type": "string"
								},
								"gt": {
									"type": "string"
								},
								"lt": {
									"type": "string"
								},
								"ne": {
									"type": "string"
								},
								"contains": {
									"type": "string"
								},
								"icontains": {
									"type": "string"
								}
							}
						}
					},
					"fields": {
						"title": "Fields",
						"type": "array",
						"items": {
							"type": "string"
						},
						"description": "A list of field names to include in the response. If not provided, all fields will be returned."
					},
					"page": {
						"title": "Page",
						"type": "integer",
						"default": 1,
						"description": "Page number for pagination. Default: 1"
					},
					"page_size": {
						"title": "Page Size",
						"type": "integer",
						"default": 100,
						"description": "Number of items per page. Default: 100"
					},
					"org_id": {
						"title": "Organization id",
						"type": "string",
						"nullable": true,
						"description": "The unique identifier of the organization associated with the memory."
					},
					"project_id": {
						"title": "Project id",
						"type": "string",
						"nullable": true,
						"description": "The unique identifier of the project associated with the memory."
					}
				}
			},
...
		},
		"securitySchemes": {
			"ApiKeyAuth": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization",
				"description": "API key authentication. Prefix your Mem0 API key with 'Token '. Example: 'Token your_api_key'"
			}
		}
	},
...
}